backend/requirements.txt
# Backend Dependencies
groq>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.4
flask>=3.0.0
pypdf2>=3.0.0
python-docx>=1.0.0
werkzeug>=3.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
pandas>=2.0.0